# analyzer: improve performance to execute tools on Docker