Previously some tools that we execute on Docker were running
synchronously. This commit fix this issue to execute all tools and
analysis in parallel. A benchmark was made using the Go examples
directory.

Before change:
goos: linux
goarch: amd64
pkg: github.com/ZupIT/horusec/internal/controllers/analyzer
cpu: Intel(R) Core(TM) i5-7200U CPU @ 2.50GHz
BenchmarkAnalyzerAnalyze
BenchmarkAnalyzerAnalyze-4                 1        33251799589 ns/op        4095008 B/op      20445 allocs/op
ok      github.com/ZupIT/horusec/internal/controllers/analyzer  33.305s

After change:
goos: linux
goarch: amd64
pkg: github.com/ZupIT/horusec/internal/controllers/analyzer
cpu: Intel(R) Core(TM) i5-7200U CPU @ 2.50GHz
BenchmarkAnalyzerAnalyze
BenchmarkAnalyzerAnalyze-4                 1        17626114733 ns/op        4089352 B/op      19347 allocs/op
ok      github.com/ZupIT/horusec/internal/controllers/analyzer  17.677s

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>

<!--
Customized from the template (https://github.com/docker/cli/blob/master/.github/PULL_REQUEST_TEMPLATE.md)

Please make sure you've read and understood our contributing guidelines;
https://github.com/ZupIT/horusec/blob/master/CONTRIBUTING.md

Please provide the following information:
-->

**- What I did**

**- How to verify it**

**- Description for the changelog**
<!--
Write a short (one line) summary that describes the changes in this
pull request for inclusion in the changelog:
-->